requirements.txt
pytesseract
Pillow
opencv-python
opencv-python-headless
